acceptcommand works as expected

+++ src/main/java/wh1spr/bot/commands/maelstrom/AcceptCommand.java
@@ -21,12 +21,18 @@ public class AcceptCommand extends Command {
 	}
 	
 	private static final String welcomeId = "433718059254153216";
-	private static MaelstromModule mael = Database2.getMaelstrom();
+	private static MaelstromModule mael = null;
 
 	@Override
 	public void onCall(JDA jda, Guild guild, MessageChannel channel, User invoker, Message message, List<String> args) {
 		if (Perm.has(Perm.TRIAL, invoker)) return; // no need for this command
 		if (!channel.getId().equals("433702216856240138")) return; // wrong channel and/or guild
+		
+		if (mael == null) mael = Database2.getMaelstrom();
+		if (!mael.isReady()) {
+			warning(message);
+			return;
+		}
 		
 		if (mael.accepted(invoker)) {
 			message.delete().queue();
